Check unset variables and unmounted volumes in the entrypoint

+++ entrypoint.sh
@@ -1,10 +1,52 @@
 #!/bin/sh
 
+set -e
+
+stop=0
+
+# Check required environment variables.
+[ "$DAEMON_UID"       == "" ] && echo "[!!!] Please set PHP_UID in your .env file." && stop=1
+[ "$DAEMON_GID"       == "" ] && echo "[!!!] Please set PHP_GID in your .env file." && stop=1
+[ "$WP_ENV"           == "" ] && echo "[!!!] Please set WP_ENV in your .env file." && stop=1 || \
+if [ "$WP_ENV" != development ] && [ "$WP_ENV" != production ]; then
+  echo "[!!!] Invalid value for WP_ENV. Please set WP_ENV to either 'development' or 'production' in your .env file."
+  stop=1
+fi
+[ "$WP_HOME"          == "" ] && echo "[!!!] Please set WP_HOME in your .env file." && stop=1
+[ "$WP_SITEURL"       == "" ] && echo "[!!!] Please set WP_SITEURL in your .env file." && stop=1
+[ "$DB_USER"          == "" ] && echo "[!!!] Please set DB_USERNAME in your .env file." && stop=1
+[ "$DB_PASSWORD"      == "" ] && echo "[!!!] Please set DB_PASSWORD in your .env file." && stop=1
+[ "$DB_NAME"          == "" ] && echo "[!!!] Please set DB_DATABASE in your .env file." && stop=1
+[ "$DB_CHARSET"       == "" ] && echo "[!!!] Please set DB_CHARSET in your .env file." && stop=1
+[ "$DB_COLLATE"       == "" ] && echo "[!!!] Please set DB_COLLATE in your .env file." && stop=1
+[ "$DB_TABLE_PREFIX"  == "" ] && echo "[!!!] Please set DB_TABLE_PREFIX in your .env file." && stop=1
+[ "$AUTH_KEY"         == "" ] && echo "[!!!] Please set AUTH_KEY in your .env file." && stop=1
+[ "$SECURE_AUTH_KEY"  == "" ] && echo "[!!!] Please set SECURE_AUTH_KEY in your .env file." && stop=1
+[ "$LOGGED_IN_KEY"    == "" ] && echo "[!!!] Please set LOGGED_IN_KEY in your .env file." && stop=1
+[ "$NONCE_KEY"        == "" ] && echo "[!!!] Please set NONCE_KEY in your .env file." && stop=1
+[ "$AUTH_SALT"        == "" ] && echo "[!!!] Please set AUTH_SALT in your .env file." && stop=1
+[ "$SECURE_AUTH_SALT" == "" ] && echo "[!!!] Please set SECURE_AUTH_SALT in your .env file"
+[ "$LOGGED_IN_SALT"   == "" ] && echo "[!!!] Please set LOGGED_IN_SALT in your .env file." && stop=1
+[ "$NONCE_SALT"       == "" ] && echo "[!!!] Please set NONCE_SALT in your .env file." && stop=1
+
+# Check directories.
+for d in wp-content wp-content/plugins wp-content/themes wp-content/uploads; do
+  if ! test -d "/var/www/html/$d"; then
+    echo "[!!!] Directory /var/www/html/wp-content/$d/ not found. Please mount it in your docker-compose.yml"
+    stop=1
+  fi
+done
+
+# Stop if anything is wrong.
+[ "$stop" == '1' ] && exit 1; unset stop
+
+# Set the UID and GID for the daemon
 sed -Ei -e '/^www-data:/{s,:82:82:,:'${DAEMON_UID:-1000}:${DAEMON_GID:-1000}':,}' /etc/passwd*
 sed -Ei -e '/^www-data:/{s,:82:,:'${DAEMON_GID:-1000}':,g}' /etc/group*
 
+# Do not allow a wp-config.php inside /var/www/html/
 if [ -f /var/www/html/wp-config.php ]; then
-  echo "[!!!] '/var/www/html/wp-config.php' removed, please set your configuration in .env"
+  echo "[!!!] /var/www/html/wp-config.php removed. Your configuration should be set in the .env file"
   rm /var/www/html/wp-config.php
 fi
 
@@ -15,9 +57,10 @@ test -e /var/log/php-fpm/error.log  || touch /var/log/php-fpm/error.log
 # Change owner and group of all writable files to www-data:www-data
 chown -R www-data:www-data /var/www/html/ /var/log/php-fpm/ /var/cache/php-opcache/
 
-echo '<?php' > /var/www/wp-config.php
-echo "define( 'DB_HOST', 'mariadb' );" >> /var/www/wp-config.php
-echo "define( 'DISALLOW_FILE_EDIT', true );" >> /var/www/wp-config.php
+# Create /var/www/wp-config.php from the environment variables
+echo '<?php' >/var/www/wp-config.php
+echo "define( 'DB_HOST', 'mariadb' );" >>/var/www/wp-config.php
+echo "define( 'DISALLOW_FILE_EDIT', true );" >>/var/www/wp-config.php
 vars="
   WP_ENV
   WP_HOME
@@ -52,29 +95,28 @@ unset DB_TABLE_PREFIX
 # Silence is golden.
 for d in wp-content wp-content/plugins wp-content/themes wp-content/uploads; do
   d="/var/www/html/$d"
-  if [ ! -f "$d/index.php" ]; then
-    echo -e '<?php\n// Silence is golden.' > "$d/index.php"
-  else
-    if [ "$(sha1sum "$d/index.php" | cut -d' ' -f1)" != "805d36c17119a980e4358437b821d369e1314816" ]; then
-      echo "[!!!] '$d/index.php' has been changed!"
-      echo -e '<?php\n// Silence is golden.' > "$d/index.php"
+  if [ -f "$d/index.php" ]; then
+    if [ "$(sha256sum "$d/index.php" | cut -d' ' -f1)" == "bd48c24ce4500e60d2524571756787283f0d3fbfa50b116331309c09773f0cd1" ]; then
+      continue
+    else
+      echo "[!!!] '$d/index.php' has changed!"
     fi
   fi
+  echo -e '<?php\n// Silence is golden.' >"$d/index.php"
 done
 
 if [ "${1#-}" != "$1" ]; then set -- php-fpm "$@"; fi
 
 if [ "$1" = 'php-fpm' ]; then
-  while ! nc -z mariadb 3306; do
-    sleep 0.1
-  done
   wp core verify-checksums || {
-    echo "[!!!] Checksums of Wordpress core files are not valid."
-    echo "[!!!] please run 'wp core verify-checksums' manually."
+    echo "[!!!] Checksums of Wordpress core files are not valid. Please run 'wp core verify-checksums' manually."
+    exit 1
   }
+  while ! nc -z mariadb 3306; do
+    sleep 0.2
+  done
   wp plugin verify-checksums --all || {
-    echo "[!!!] Checksums of Wordpress plugins are not valid."
-    echo "[!!!] please run 'wp plugin verify-checksums --all' manually."
+    echo "[!!!] Checksums of Wordpress plugins are not valid. Please run 'wp plugin verify-checksums --all' manually."
   }
 fi
 exec "$@"
